+++ flink-table-store-connector/src/test/java/org/apache/flink/table/store/connector/source/TestChangelogDataReadWrite.java
@@ -30,7 +30,6 @@ import org.apache.flink.table.store.file.io.DataFileMeta;
 import org.apache.flink.table.store.file.memory.HeapMemorySegmentPool;
 import org.apache.flink.table.store.file.memory.MemoryOwner;
 import org.apache.flink.table.store.file.mergetree.compact.DeduplicateMergeFunction;
-import org.apache.flink.table.store.file.mergetree.compact.MergeFunctionFactory;
 import org.apache.flink.table.store.file.operation.KeyValueFileStoreRead;
 import org.apache.flink.table.store.file.operation.KeyValueFileStoreWrite;
 import org.apache.flink.table.store.file.predicate.Predicate;
@@ -109,7 +108,7 @@ public class TestChangelogDataReadWrite {
                         KEY_TYPE,
                         VALUE_TYPE,
                         COMPARATOR,
-                        MergeFunctionFactory.of(new DeduplicateMergeFunction()),
+                        DeduplicateMergeFunction.factory(),
                         avro,
                         pathFactory);
         return new KeyValueTableRead(read) {
@@ -160,7 +159,7 @@ public class TestChangelogDataReadWrite {
                                 KEY_TYPE,
                                 VALUE_TYPE,
                                 () -> COMPARATOR,
-                                new DeduplicateMergeFunction(),
+                                DeduplicateMergeFunction.factory(),
                                 pathFactory,
                                 snapshotManager,
                                 null, // not used, we only create an empty writer
