+++ flink-table-store-core/src/main/java/org/apache/flink/table/store/file/KeyValueFileStore.java
@@ -90,7 +90,7 @@ public class KeyValueFileStore extends AbstractFileStore<KeyValue> {
                 keyType,
                 valueType,
                 keyComparatorSupplier,
-                mfFactory.create(null),
+                mfFactory,
                 pathFactory(),
                 snapshotManager(),
                 newScan(true),
