+++ flink-table-store-core/src/main/java/org/apache/flink/table/store/file/operation/KeyValueFileStoreWrite.java
@@ -33,7 +33,7 @@ import org.apache.flink.table.store.file.mergetree.MergeTreeWriter;
 import org.apache.flink.table.store.file.mergetree.compact.CompactRewriter;
 import org.apache.flink.table.store.file.mergetree.compact.CompactStrategy;
 import org.apache.flink.table.store.file.mergetree.compact.FullChangelogMergeTreeCompactRewriter;
-import org.apache.flink.table.store.file.mergetree.compact.MergeFunction;
+import org.apache.flink.table.store.file.mergetree.compact.MergeFunctionFactory;
 import org.apache.flink.table.store.file.mergetree.compact.MergeTreeCompactManager;
 import org.apache.flink.table.store.file.mergetree.compact.MergeTreeCompactRewriter;
 import org.apache.flink.table.store.file.mergetree.compact.UniversalCompaction;
@@ -64,7 +64,7 @@ public class KeyValueFileStoreWrite extends MemoryFileStoreWrite<KeyValue> {
     private final KeyValueFileReaderFactory.Builder readerFactoryBuilder;
     private final KeyValueFileWriterFactory.Builder writerFactoryBuilder;
     private final Supplier<Comparator<RowData>> keyComparatorSupplier;
-    private final MergeFunction<KeyValue> mergeFunction;
+    private final MergeFunctionFactory<KeyValue> mfFactory;
     private final CoreOptions options;
     private final FileStorePathFactory pathFactory;
 
@@ -75,7 +75,7 @@ public class KeyValueFileStoreWrite extends MemoryFileStoreWrite<KeyValue> {
             RowType keyType,
             RowType valueType,
             Supplier<Comparator<RowData>> keyComparatorSupplier,
-            MergeFunction<KeyValue> mergeFunction,
+            MergeFunctionFactory<KeyValue> mfFactory,
             FileStorePathFactory pathFactory,
             SnapshotManager snapshotManager,
             FileStoreScan scan,
@@ -98,7 +98,7 @@ public class KeyValueFileStoreWrite extends MemoryFileStoreWrite<KeyValue> {
                         pathFactory,
                         options.targetFileSize());
         this.keyComparatorSupplier = keyComparatorSupplier;
-        this.mergeFunction = mergeFunction;
+        this.mfFactory = mfFactory;
         this.options = options;
         this.pathFactory = pathFactory;
     }
@@ -151,7 +151,7 @@ public class KeyValueFileStoreWrite extends MemoryFileStoreWrite<KeyValue> {
                 compactManager,
                 getMaxSequenceNumber(restoreFiles),
                 keyComparator,
-                mergeFunction.copy(),
+                mfFactory.create(null),
                 writerFactory,
                 options.commitForceCompact(),
                 options.changelogProducer());
@@ -199,10 +199,10 @@ public class KeyValueFileStoreWrite extends MemoryFileStoreWrite<KeyValue> {
                     readerFactory,
                     writerFactory,
                     keyComparator,
-                    mergeFunction);
+                    mfFactory);
         } else {
             return new MergeTreeCompactRewriter(
-                    readerFactory, writerFactory, keyComparator, mergeFunction);
+                    readerFactory, writerFactory, keyComparator, mfFactory);
         }
     }
 }
